moved routeConfig from shared to app layer to comply with fsd layers rules. Now app layer imports pages layer. Before it was pages layed being imported into shared
