+++ src/app/providers/router/ui/RequireAuth.tsx
@@ -1,7 +1,7 @@
 import { useSelector } from 'react-redux';
 import { UserRole, getUserAuthData, getUserRoles } from '@/entities/User';
 import { Navigate, useLocation } from 'react-router-dom';
-import { RoutePath } from '@/shared/config/routeConfig/routeConfig';
+import { RoutePath } from '@/shared/const/router';
 import { useMemo } from 'react';
 
 interface RequireAuthProps {
