+++ src/features/AvatarDropdown/ui/AvatarDropdown.tsx
@@ -1,7 +1,7 @@
 import { FC, memo, useCallback } from 'react';
 import { classNames } from '@/shared/lib/classnames/classNames';
 import { useTranslation } from 'react-i18next';
-import { RoutePath } from '@/shared/config/routeConfig/routeConfig';
+import { RoutePath } from '@/shared/const/router';
 import { Avatar } from '@/shared/ui/Avatar/Avatar';
 import { Dropdown } from '@/shared/ui/Popups';
 import {
