+++ src/shared/types/router.ts
@@ -1,0 +1,10 @@
+import { UserRole } from "@/entities/User";
+import { RouteProps } from "react-router-dom";
+
+
+
+
+export type AppRoutesProps = RouteProps & {
+    authOnly?: boolean;
+    roles?: UserRole[];
+};
