+++ src/widgets/Sidebar/model/selectors/getSidebarItems.ts
@@ -1,6 +1,6 @@
 import ArticleIcon from '@/shared/assets/icons/article-20-20.svg';
 import ProfileIcon from '@/shared/assets/icons/profile-20-20.svg';
-import { RoutePath } from '@/shared/config/routeConfig/routeConfig';
+import { RoutePath } from '@/shared/const/router';
 import AboutIcon from '@/shared/assets/icons/about-20-20.svg';
 import MainIcon from '@/shared/assets/icons/main-20-20.svg';
 import { SidebarItemType } from '../types/sidebar';
